refactor: move MessageParseArgs to its own file
